requirements.txt
absl-py 
aiohttp 
aiosignal 
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
astunparse==1.6.3
async-timeout 
attrs 
beautifulsoup4==4.12.3
blinker 
Brotli 
cachetools 
certifi 
cffi 
charset-normalizer 
click 
colorama 
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.1.1
cryptography 
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
flatbuffers==2.0.7
fonttools==4.53.1
frozenlist 
gast 
google-ai-generativelanguage==0.1.0
google-api-core==2.19.1
google-auth 
google-auth-oauthlib 
google-generativeai==0.1.0rc1
google-pasta 
googleapis-common-protos==1.63.2
grpcio 
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna 
importlib-metadata 
importlib_resources==6.4.0
Jinja2==3.1.4
keras==2.10.0
Keras-Applications 
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown 
markdown-it-py==3.0.0
MarkupSafe 
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft 
mkl-random 
mkl-service==2.4.0
mltu==0.1.4
mpmath==1.3.0
multidict 
numpy 
oauthlib 
onnx==1.12.0
onnxruntime==1.18.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum 
packaging 
pandas==2.0.3
pillow==10.4.0
platformdirs 
pooch 
proto-plus==1.24.0
protobuf==4.25.3
pyasn1 
pyasn1-modules==0.2.8
pycparser 
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT 
pyOpenSSL 
pyparsing==3.1.2
pyreadline==2.1
pyreadline3==3.4.1
PySocks 
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests 
requests-oauthlib 
rich==13.7.1
rsa 
scipy==1.10.1
shellingham==1.5.4
six 
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
stow==1.4.1
sympy==1.13.1
tensorboard 
tensorboard-data-server 
tensorboard-plugin-wit 
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor 
tf2onnx==1.14.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3 
uvicorn==0.30.4
watchfiles==0.22.0
websockets==12.0
Werkzeug 
win-inet-pton 
wrapt 
yarl 
zipp 
